Refactored providers. Introduced abstract classes to improve code readability.
